[pentest] Split firmware into SCA & FI
Due to the rising number of tests added to the pen testing framework,
the combined image exceeds the available IMEM. Hence, this commit
splits up the binary into a separate FI and SCA binary for the chip.
On the FPGA, still the combined image can be used.
