+++ sw/device/tests/penetrationtests/firmware/fi/crypto_fi.c
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "sw/device/tests/penetrationtests/firmware/crypto_fi.h"
+#include "sw/device/tests/penetrationtests/firmware/fi/crypto_fi.h"
 
 #include "hw/ip/aes/model/aes_modes.h"
 #include "sw/device/lib/base/memory.h"
@@ -15,7 +15,7 @@
 #include "sw/device/lib/testing/test_framework/ujson_ottf.h"
 #include "sw/device/lib/ujson/ujson.h"
 #include "sw/device/sca/lib/sca.h"
-#include "sw/device/tests/penetrationtests/firmware/sca_lib.h"
+#include "sw/device/tests/penetrationtests/firmware/lib/sca_lib.h"
 #include "sw/device/tests/penetrationtests/json/crypto_fi_commands.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
