+++ sw/device/tests/penetrationtests/firmware/fi/ibex_fi.S
