+++ sw/device/tests/penetrationtests/firmware/fi/otbn/BUILD
