+++ sw/device/tests/penetrationtests/firmware/fi/otbn/otbn_char_hardware_dmem_op_loop.s
