+++ sw/device/tests/penetrationtests/firmware/fi/otbn/otbn_char_hardware_reg_op_loop.s
