+++ sw/device/tests/penetrationtests/firmware/fi/otbn/otbn_char_unrolled_dmem_op_loop.s
