+++ sw/device/tests/penetrationtests/firmware/fi/otbn/otbn_char_unrolled_reg_op_loop.s
