+++ sw/device/tests/penetrationtests/firmware/fi/otbn/otbn_key_sideload.s
