+++ sw/device/tests/penetrationtests/firmware/fi/otbn/otbn_load_integrity.s
